chore: 폴더 구조를 기능별로 구분
