--- src/components/FrameSelect/PreviewFrame.tsx
@@ -1,11 +1,0 @@
-import { PhotoFrame } from '@/components/Photo';
-
-const PreviewFrame = () => {
-  return (
-    <div className="flex flex-col items-center gap-3 p-3">
-      <PhotoFrame />
-    </div>
-  );
-};
-
-export default PreviewFrame;
