+++ src/components/camera/index.ts
@@ -1,0 +1,1 @@
+export { default as CameraSection } from './CameraSection';
