+++ src/components/filter/index.ts
@@ -1,0 +1,1 @@
+export { default as FilterSelectSection } from './FilterSelectSection';
