+++ src/components/frame/index.ts
@@ -1,2 +1,1 @@
 export { default as FrameSelectSection } from './FrameSelectSection';
-export { default as PreviewFrame } from './PreviewFrame';
