+++ src/components/preview/PreviewSection.tsx
@@ -1,0 +1,11 @@
+import { ToastFrame } from '@/components/preview';
+
+const PreviewSection = () => {
+  return (
+    <section className="h-full w-[300px] flex flex-col justify-center items-center">
+      <ToastFrame />
+    </section>
+  );
+};
+
+export default PreviewSection;
