--- src/components/takePhoto/CameraFrame.tsx
@@ -1,11 +1,0 @@
-import { ToastFrame } from '@/components/takePhoto';
-
-const CameraFrame = () => {
-  return (
-    <div className="h-full w-[300px] flex flex-col justify-center items-center">
-      <ToastFrame />
-    </div>
-  );
-};
-
-export default CameraFrame;
