--- src/components/takePhoto/index.ts
@@ -1,5 +1,0 @@
-export { default as CameraCapture } from './CameraCapture';
-export { default as CameraFrame } from './CameraFrame';
-export { default as FilterSelectSection } from './FilterSelectSection';
-export { default as ToastFrame } from './ToastFrame';
-export { default as ToastFrameCut } from './ToastFrameCut';
