+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-export { default as FrameSelectPage } from './FrameSelectPage';
+export { default as CustomPhotoPage } from './CustomPhotoPage';
 export { default as SavePhotoPage } from './SavePhotoPage';
 export { default as TakePhotoPage } from './TakePhotoPage';
